구조 변경에 따른 Product 관련 컨트롤러, 매퍼, 서비스 로직 변경

+++ src/main/java/com/joo/abysshop/dto/product/ProductDetailResponse.java
@@ -10,9 +10,9 @@ import org.springframework.web.multipart.MultipartFile;
 @AllArgsConstructor
 public class ProductDetailResponse {
 
-    private Integer productId;
+    private Long productId;
     private String productName;
-    private Integer price;
+    private Long price;
     private String description;
     private MultipartFile image;
 }
